requirements.txt
# 환경변수 관리
python-dotenv>=1.0.0

# 브라우저 자동화
playwright>=1.40.0

# HTTP 클라이언트
requests>=2.31.0

# Gemini API (새로운 통합 SDK - 2025)
google-genai>=0.6.0

# 이미지 처리
Pillow>=10.0.0

# 데이터 처리
pandas>=2.0.0
